The previous approach will not work. New approach.
We can't handle $ signs without escaping them.
We can't take multiple sets of expressions, such as
  verb1|verb2|verb3 verba|verbb|verbc should = verb1 verbc or
  something like that, because of the space.

Can't handle executing commands because the poem is
processed by searching for a pair of angle brackets, and
detecting a command would immediatly cause the program to
exit when we really want it to process all the text up to
the command. Ex : verb1|verb2|verb3|$END means we'll just
exit immediatly and not consider a random selection of those
verbs.